fix the build, sorry for the long delay
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@17922 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/network/compat/libnet/socket.c
@@ -313,7 +313,7 @@ _EXPORT ssize_t send(int sock, const void *data, size_t datalen, int flags)
 }
 
 
-_EXPORT int getsockopt(int sock, int level, int option, void *optval, size_t *optlen)
+_EXPORT int getsockopt(int sock, int level, int option, void *optval, int *optlen)
 {
 	struct stack_driver_args args;
 	int rv;
@@ -341,7 +341,7 @@ _EXPORT int getsockopt(int sock, int level, int option, void *optval, size_t *op
 	return rv;
 }
 
-_EXPORT int setsockopt(int sock, int level, int option, const void *optval, size_t optlen)
+_EXPORT int setsockopt(int sock, int level, int option, const void *optval, int optlen)
 {
 	struct stack_driver_args args;
